otterdog: set new default branch for manifest repo
The main branch does not exist in the original repo we mirror
from (and thus creates problems when running git push --mirror).

In addition, our changes will be targeted into the
OpenHarmony-3.2-Release branch, which means we should show it
as default to make is easier to discover for contributors.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>